this makefile runs, but segfaults on windows
git-svn-id: https://forge.cornell.edu/svn/repos/nlsl_f90/trunk@46 85800f7d-5aad-4a23-bede-8b4c7eff9ab2

+++ Makefile
@@ -52,19 +52,21 @@ LIBS = -L/usr/X11R6/lib -lX11 -lc
 #-----------------------------------------------------------------------
 #
 
-
-SPCI = nlsdim.inc eprprm.inc expdat.inc parcom.inc tridag.inc basis.inc pidef.inc
-BASI = nlsdim.inc eprprm.inc parcom.inc basis.inc stdio.inc
-LETI = nlsdim.inc eprprm.inc expdat.inc prmeqv.inc parcom.inc lpnam.inc
-NLSI = nlsdim.inc eprprm.inc expdat.inc lmcom.inc stdio.inc parcom.inc
-PRMI = nlsdim.inc eprprm.inc lmcom.inc parcom.inc lpnam.inc
-FUNI = nlsdim.inc eprprm.inc expdat.inc tridag.inc parcom.inc iterat.inc
-FNI2 = basis.inc pidef.inc stdio.inc prmeqv.inc mspctr.inc ftwork.inc 
-FITI = nlsnam.inc eprmat.inc errmsg.inc mspctr.inc iterat.inc
-MATI = nlsdim.inc eprprm.inc eprmat.inc rndoff.inc
-CHKI = maxl.inc nlsdim.inc eprprm.inc rndoff.inc prmeqv.inc errmsg.inc
-DATI = nlsdim.inc expdat.inc mspctr.inc nlsnam.inc stdio.inc parcom.inc 
-SCLI = nlsdim.inc expdat.inc lmcom.inc parcom.inc mspctr.inc iterat.inc 
+MOD1 = eprprm.o parcom.o nlsdim.o errmsg.o lpnam.o expdat.o basis.o
+MOD2 = symdef.o maxl.o bincom.o dfunc.o eprmat.o ftwork.o iterat.o lmcom.o mspctr.o mtsdef.o 
+MOD3 = nlsnam.o parsav.o physcn.o pidef.o rnddbl.o stdio.o timer.o tridag.o 
+SPCI = nlsdim.mod eprprm.mod expdat.mod parcom.mod tridag.mod basis.mod pidef.mod
+BASI = nlsdim.mod eprprm.mod parcom.mod basis.mod stdio.mod
+LETI = nlsdim.mod eprprm.mod expdat.mod parcom.mod lpnam.mod
+NLSI = nlsdim.mod nlsnam.mod eprprm.mod expdat.mod lmcom.mod parcom.mod symdef.mod stdio.mod parcom.mod
+PRMI = nlsdim.mod eprprm.mod lmcom.mod parcom.mod lpnam.mod
+FUNI = nlsdim.mod eprprm.mod expdat.mod tridag.mod parcom.mod iterat.mod
+FNI2 = basis.mod pidef.mod stdio.mod mspctr.mod ftwork.mod 
+FITI = nlsnam.mod eprmat.mod errmsg.mod mspctr.mod iterat.mod
+MATI = nlsdim.mod eprprm.mod eprmat.mod rnddbl.mod
+CHKI = maxl.mod nlsdim.mod eprprm.mod rnddbl.mod errmsg.mod
+DATI = nlsdim.mod expdat.mod mspctr.mod nlsnam.mod stdio.mod parcom.mod 
+SCLI = nlsdim.mod expdat.mod lmcom.mod parcom.mod mspctr.mod iterat.mod 
 CGO  = cscg.o zaypx.o cgltri.o zdotu.o zaxpy.o scmvm.o tdsqz.o tdchek.o
 EPO1 = momdls.o eprls.o matrll.o cd2km.o anxlk.o w3j.o stvect.o cfs.o $(CGO)
 EPO2 = pstvec.o pmatrl.o ccrints.o fz.o bessel.o plgndr.o ipar.o
@@ -79,13 +81,11 @@ NLSF = fitl.o lfun.o lcheck.o setspc.o $(NLSH) $(EPO1) $(EPO2) ordrpr.o
 NLSL = lmnls.o enorm.o qrfac.o lmpar.o qrsolv.o covar.o
 NLSN = dchex.o daxpy.o dcopy.o ddot.o drotg.o
 NLSP = pltx.o 
-NLSS = strutl.o lprmpt.o catch.o ipfind.o nlstxt.o  
+NLSS = strutl1.o strutl2.o lprmpt.o catch.o ipsfind.o 
 ifeq ($(OS),"windows")
-NLSO = nlsl.o pltx_dummy.o $(NLSC) $(NLSS) $(NLSD) $(NLSF) $(NLSL) $(NLSN) $(NLSB)\
-       $(NLSW)
+NLSO = nlsl.o pltx_dummy.o $(MOD1) $(MOD2) $(MOD3) $(NLSC) $(NLSS) $(NLSD) $(NLSF) $(NLSL) $(NLSN) $(NLSB) $(NLSW)
 else
-NLSO = nlsl.o $(NLSC) $(NLSS) $(NLSD) $(NLSF) $(NLSL) $(NLSP) $(NLSN) $(NLSB)\
-       $(NLSW)
+NLSO = nlsl.o $(MOD1) $(MOD2) $(MOD3) $(NLSC) $(NLSS) $(NLSD) $(NLSF) $(NLSL) $(NLSP) $(NLSN) $(NLSB) $(NLSW)
 endif
 
 #-----------------------------------------------------------------------
@@ -102,86 +102,86 @@ all             : nlsl
 clean           :
 			rm -f *.o nlsl
 endif
-addprm.o	: addprm.f nlsdim.inc eprprm.inc expdat.inc parcom.inc lpnam.inc\
-                  lmcom.inc stdio.inc rndoff.inc prmeqv.inc
-assgnc.o	: assgnc.f $(BASI)
-bessel.o	: bessel.f rndoff.inc
-basisc.o	: basisc.f $(BASI)
+addprm.o	: addprm.f90 nlsdim.mod eprprm.mod expdat.mod parcom.mod lpnam.mod\
+                  lmcom.mod stdio.mod rnddbl.mod
+assgnc.o	: assgnc.f90 $(BASI)
+bessel.o	: bessel.f90 rnddbl.mod
+basisc.o	: basisc.f90 $(BASI)
 catch.o         : catch.c fortrancall.h
-cd2km.o         : cd2km.f
-cgltri.o	: cgltri.f nlsdim.inc rndoff.inc
-confc.o		: confc.f iterat.inc stdio.inc
-convlv.o	: convlv.f nlsdim.inc
-convtc.o	: convtc.f nlsdim.inc eprprm.inc stdio.inc lpnam.inc parcom.inc
-covrpt.o	: covrpt.f nlsdim.inc eprprm.inc expdat.inc parcom.inc\
-		  lmcom.inc lpnam.inc mspctr.inc iterat.inc rndoff.inc\
-		  stdio.inc
-cscg.o		: cscg.f nlsdim.inc eprmat.inc rndoff.inc 
-cslnzs.o 	: cslnzs.f nlsdim.inc eprmat.inc
-datac.o		: datac.f $(DATI) mspctr.inc lmcom.inc rndoff.inc
-eprls.o		: eprls.f $(MATI) ftwork.inc tridag.inc pidef.inc
-fitc.o		: fitc.f nlsdim.inc lmcom.inc parcom.inc stdio.inc iterat.inc
-fitl.o  	: fitl.f $(NLSI) $(FITI) rndoff.inc timer.inc
-fixc.o		: fixc.f nlsdim.inc eprprm.inc prmeqv.inc parcom.inc lpnam.inc stdio.inc
-ftfuns.o	: ftfuns.f
-fz.o		: fz.f eprprm.inc
-gconvl.o	: gconvl.f nlsdim.inc eprprm.inc ftwork.inc pidef.inc
+cd2km.o         : cd2km.f90
+cgltri.o	: cgltri.f90 nlsdim.mod rnddbl.mod
+confc.o		: confc.f90 iterat.mod stdio.mod
+convlv.o	: convlv.f90 nlsdim.mod
+convtc.o	: convtc.f90 nlsdim.mod eprprm.mod stdio.mod lpnam.mod parcom.mod
+covrpt.o	: covrpt.f90 nlsdim.mod eprprm.mod expdat.mod parcom.mod\
+		  lmcom.mod lpnam.mod mspctr.mod iterat.mod rnddbl.mod\
+		  stdio.mod
+cscg.o		: cscg.f90 nlsdim.mod eprmat.mod rnddbl.mod 
+cslnzs.o 	: cslnzs.f90 nlsdim.mod eprmat.mod
+datac.o		: datac.f90 $(DATI) mspctr.mod lmcom.mod rnddbl.mod
+eprls.o		: eprls.f90 $(MATI) ftwork.mod tridag.mod pidef.mod
+fitc.o		: fitc.f90 nlsdim.mod lmcom.mod parcom.mod stdio.mod iterat.mod
+fitl.o  	: fitl.f90 $(NLSI) $(FITI) rnddbl.mod timer.mod
+fixc.o		: fixc.f90 nlsdim.mod eprprm.mod parcom.mod lpnam.mod stdio.mod
+ftfuns.o	: ftfuns.f90
+fz.o		: fz.f90 eprprm.mod
+gconvl.o	: gconvl.f90 nlsdim.mod eprprm.mod ftwork.mod pidef.mod
 genio.o         : genio.c fortrancall.h
-getids.o	: getids.f nlsdim.inc stdio.inc nlsnam.inc
-getdat.o	: getdat.f nlsdim.inc stdio.inc
-helpc.o		: helpc.f stdio.inc
-ipfind.o	: ipfind.f $(BASI) eprprm.inc parcom.inc lpnam.inc
-lbasix.o        : lbasix.f nlsdim.inc eprprm.inc mtsdef.inc stdio.inc
-lcheck.o	: lcheck.f $(CHKI) prmeqv.inc
-letc.o		: $(LETI) stdio.inc
-lfun.o		: lfun.f $(FUNI) $(FNI2) 
-lgrint.o	: lgrint.f
-lmnls.o		: lmnls.f iterat.inc
+getids.o	: getids.f90 nlsdim.mod stdio.mod nlsnam.mod
+getdat.o	: getdat.f90 nlsdim.mod stdio.mod
+helpc.o		: helpc.f90 stdio.mod
+ipsfind.o 	: ipsfind.f90 $(BASI) lpnam.mod expdat.mod
+lbasix.o        : lbasix.f90 nlsdim.mod eprprm.mod mtsdef.mod stdio.mod
+lcheck.o	: lcheck.f90 $(CHKI)
+letc.o		: $(LETI) stdio.mod
+lfun.o		: lfun.f90 $(FUNI) $(FNI2) 
+lgrint.o	: lgrint.f90
+lmnls.o		: lmnls.f90 iterat.mod
 lprmpt.o        : lprmpt.c fortrancall.h
-l1pfun.o	: l1pfun.f nlsdim.inc expdat.inc parcom.inc
-matrll.o	: matrll.f $(MATI) maxl.inc physcn.inc
-mnbrak.o	: mnbrak.f errmsg.inc
-momdls.o	: momdls.f $(MATI) pidef.inc stdio.inc dfunc.inc
-nlsl.o		: nlsl.f $(NLSI) parcom.inc tridag.inc basis.inc iterat.inc
-nlstxt.o	: nlstxt.f nlsdim.inc eprprm.inc lpnam.inc errmsg.inc
-ordrpr.o	: ordrpr.f rndoff.inc dfunc.inc pidef.inc
-parc.o		: parc.f nlsdim.inc eprprm.inc expdat.inc parcom.inc lpnam.inc stdio.inc\
-                  rndoff.inc symdef.inc mtsdef.inc
+l1pfun.o	: l1pfun.f90 nlsdim.mod expdat.mod parcom.mod
+matrll.o	: matrll.f90 $(MATI) maxl.mod physcn.mod
+mnbrak.o	: mnbrak.f90 errmsg.mod
+momdls.o	: momdls.f90 $(MATI) pidef.mod stdio.mod dfunc.mod
+nlsl.o		: nlsl.f90 $(NLSI) basis.mod iterat.mod mspctr.mod tridag.mod
+#nlstxt.o	: nlstxt.f90 nlsdim.mod eprprm.mod lpnam.mod errmsg.mod
+ordrpr.o	: ordrpr.f90 rnddbl.mod dfunc.mod pidef.mod
+parc.o		: parc.f90 nlsdim.mod eprprm.mod expdat.mod parcom.mod lpnam.mod stdio.mod\
+                  rnddbl.mod symdef.mod mtsdef.mod
 pltx.o		: pltx.c fortrancall.h
-pltx_dummy.o		: pltx_dummy.c fortrancall.h
-pmatrl.o	: pmatrl.f $(MATI) maxl.inc physcn.inc
-pstvec.o 	: pstvec.f nlsdim.inc eprprm.inc errmsg.inc rndoff.inc
-rdpar.o		: rdpar.f eprprm.inc stdio.inc
-rms.o		: rms.f nlsdim.inc mspctr.inc parcom.inc rndoff.inc
-scalec.o        : scalec.f $(SCLI) stdio.inc
-scmvm.o 	: scmvm.f nlsdim.inc rndoff.inc eprmat.inc eprprm.inc stdio.inc
-series.o	: series.f nlsdim.inc eprprm.inc expdat.inc mspctr.inc parcom.inc stdio.inc
-setmts.o	: setmts.f nlsdim.inc eprprm.inc maxl.inc rndoff.inc mtsdef.inc
-setnm.o		: setnm.f nlsdim.inc nlsnam.inc
-setprm.o	: setprm.f $(BASI) $(PRMI) iterat.inc stdio.inc symdef.inc
-setspc.o	: setspc.f $(SPCI) iterat.inc
-shiftc.o	: shiftc.f nlsdim.inc expdat.inc parcom.inc stdio.inc
-sitec.o		: sitec.f nlsdim.inc expdat.inc parcom.inc tridag.inc basis.inc stdio.inc
-srchc.o 	: srchc.f nlsdim.inc eprprm.inc lmcom.inc parcom.inc\
-                  iterat.inc errmsg.inc lpnam.inc stdio.inc expdat.inc
-sscale.o	: sscale.f nlsdim.inc
-statc.o		: statc.f $(BASI) nlsdim.inc eprprm.inc expdat.inc parcom.inc\
-                  lmcom.inc mtsdef.inc
-stats.o		: stats.f
-strutl.o	: strutl.f stdio.inc
-stvect.o 	: stvect.f nlsdim.inc eprprm.inc errmsg.inc rndoff.inc
-tdchek.o        : tdchek.f $(BASI) tridag.inc errmsg.inc mtsdef.inc
-tdqz.o		: tdsqz.f nlsdim.inc expdat.inc tridag.inc
-varyc.o		: varyc.f nlsdim.inc eprprm.inc prmeqv.inc parcom.inc lpnam.inc stdio.inc
-writr.o		: writr.f nlsdim.inc expdat.inc lmcom.inc parcom.inc stdio.inc\
-                  tridag.inc mspctr.inc iterat.inc
-writec.o	: writec.f nlsdim.inc expdat.inc parcom.inc lmcom.inc mspctr.inc stdio.inc
-w3j.o		: w3j.f maxl.inc bincom.inc stdio.inc
-zaxpy.o 	: zaxpy.f nlsdim.inc rndoff.inc
-qzaypx.o 	: zaypx.f nlsdim.inc rndoff.inc
-znormu.o	: znormu.f nlsdim.inc rndoff.inc
-zscsw.o		: zscsw.f nlsdim.inc rndoff.inc
-zdotu.o		: zdotu.f nlsdim.inc rndoff.inc
+pmatrl.o	: pmatrl.f90 $(MATI) maxl.mod physcn.mod
+pstvec.o 	: pstvec.f90 nlsdim.mod eprprm.mod errmsg.mod rnddbl.mod
+rdpar.o		: rdpar.f90 eprprm.mod stdio.mod
+rms.o		: rms.f90 nlsdim.mod mspctr.mod parcom.mod rnddbl.mod
+scalec.o        : scalec.f90 $(SCLI) stdio.mod
+scmvm.o 	: scmvm.f90 nlsdim.mod rnddbl.mod eprmat.mod eprprm.mod stdio.mod
+series.o	: series.f90 nlsdim.mod eprprm.mod expdat.mod mspctr.mod parcom.mod stdio.mod
+setmts.o	: setmts.f90 nlsdim.mod eprprm.mod maxl.mod rnddbl.mod mtsdef.mod
+setnm.o		: setnm.f90 nlsdim.mod nlsnam.mod
+setprm.o	: setprm.f90 $(BASI) $(PRMI) iterat.mod stdio.mod symdef.mod
+setspc.o	: setspc.f90 $(SPCI) iterat.mod
+shiftc.o	: shiftc.f90 nlsdim.mod expdat.mod parcom.mod stdio.mod
+sitec.o		: sitec.f90 nlsdim.mod expdat.mod parcom.mod tridag.mod basis.mod stdio.mod
+srchc.o 	: srchc.f90 nlsdim.mod eprprm.mod lmcom.mod parcom.mod\
+                  iterat.mod errmsg.mod lpnam.mod stdio.mod expdat.mod
+sscale.o	: sscale.f90 nlsdim.mod
+statc.o		: statc.f90 $(BASI) nlsdim.mod eprprm.mod expdat.mod parcom.mod\
+                  lmcom.mod mtsdef.mod errmsg.mod
+stats.o		: stats.f90
+strutl1.o	: strutl1.f90 stdio.mod
+strutl2.o	: strutl2.f90
+stvect.o 	: stvect.f90 nlsdim.mod eprprm.mod errmsg.mod rnddbl.mod
+tdchek.o        : tdchek.f90 $(BASI) tridag.mod errmsg.mod mtsdef.mod
+tdqz.o		: tdsqz.f90 nlsdim.mod expdat.mod tridag.mod stdio.mod
+varyc.o		: varyc.f90 nlsdim.mod eprprm.mod parcom.mod lpnam.mod stdio.mod
+writr.o		: writr.f90 nlsdim.mod expdat.mod lmcom.mod parcom.mod stdio.mod\
+                  tridag.mod mspctr.mod iterat.mod
+writec.o	: writec.f90 nlsdim.mod expdat.mod parcom.mod lmcom.mod mspctr.mod stdio.mod
+w3j.o		: w3j.f90 maxl.mod stdio.mod bincom.mod
+zaxpy.o 	: zaxpy.f90 nlsdim.mod rnddbl.mod
+qzaypx.o 	: zaypx.f90 nlsdim.mod rnddbl.mod
+znormu.o	: znormu.f90 nlsdim.mod rnddbl.mod
+zscsw.o		: zscsw.f90 nlsdim.mod rnddbl.mod
+zdotu.o		: zdotu.f90 nlsdim.mod rnddbl.mod
 nlsdim.o nlsdim.mod: nlsdim.f90
 parcom.o parcom.mod: parcom.f90 nlsdim.mod
 eprprm.o eprprm.mod: eprprm.f90 parcom.mod nlsdim.mod
@@ -190,9 +190,6 @@ lpnam.o lpnam.mod: lpnam.f90 nlsdim.mod
 expdat.o expdat.mod: expdat.f90 nlsdim.mod
 basis.o basis.mod: basis.f90 nlsdim.mod
 stdio.o stdio.mod: stdio.f90
-ipsfind.o: ipsfind.f90 nlsdim.mod parcom.mod eprprm.mod lpnam.mod expdat.mod basis.mod stdio.mod
-strutl1.o: strutl1.f90 stdio.mod
-strutl2.o: strutl2.f90
 maxl.o maxl.mod: maxl.f90
 bincom.o bincom.mod: bincom.f90 maxl.mod
 eprmat.o eprmat.mod: eprmat.f90 nlsdim.mod
@@ -226,7 +223,7 @@ clean:
 	$(RM) *.o *.mod nlsl testmods
 
 TESTS = testmods.o eprprm.o parcom.o nlsdim.o errmsg.o lpnam.o ipsfind.o expdat.o basis.o strutl2.o
-EXTRAS = strutl1.o symdef.o maxl.o bincom.o dfunc.o eprmat.o ftwork.o iterat.o lmcom.o mspctr.o mtsdef.o nlsnam.o parsav.o physcn.o pidef.o rndoff.o timer.o tridag.o
+EXTRAS = strutl1.o symdef.o maxl.o bincom.o dfunc.o eprmat.o ftwork.o iterat.o lmcom.o mspctr.o mtsdef.o nlsnam.o parsav.o physcn.o pidef.o rnddbl.o timer.o tridag.o
 # no tests yet for $(EXTRAS)... compile them anyway
 
 testmods: $(TESTS) $(EXTRAS)
@@ -242,9 +239,6 @@ testclean:
 %.o : %.c
 	$(CC) $(CFLAGS) $*.c
 
-%.o : %.f
-	$(F77) $(FFLAGS) $*.f
-
 %.o %.mod : %.f90
-	rm -f $*.mod
+	if [ -a $*.mod ] ; then rm -f $*.mod ; fi
 	$(F90) $(FFLAGS) -ffixed-form $*.f90
